requirements.txt
sqlalchemy~=2.0.25
fastapi~=0.88.0
jinja2~=3.1.2
uvicorn~=0.22.0
python-dotenv~=1.0.1
pydantic~=1.10.9
passlib~=1.7.4
